requirements.txt
streamlit
matplotlib
google.generativeai
python-dotenv
